mvn test

+++ vars/dockerImageScan.groovy
@@ -8,7 +8,8 @@
 def call(String aws_account_id, String region, String ecr_repoName){
     
     sh """
-    rm scan*
+    if [ -e scan* ]; then
+        rm scan*
     trivy image ${aws_account_id}.dkr.ecr.${region}.amazonaws.com/${ecr_repoName}:latest > scan-${BUILD_NUMBER}.txt
     cat scan-${BUILD_NUMBER}.txt
     """
